fix: clean up event handlers and CSS
Need to change hover to mousemove from mouseover, or just having your
mouse parked over the list as soon as it appears will select an entry.

Just rely on .highlight, otherwise two separate items will be
highlighted, both .highlight and :hover.

Clean up ^-n handler to not be so repetitive.