feat: use final image load event
Previously we were just using the initial image generation API response
to stop the loading spinner and other things, however, the second step
is to actually fetch the image file. We can use the @load handler on the
image to be notified of when the image loading is fully complete.