build: run vue-tsc and tsc separately
Conceptually, it can't run on files individually (though I would be
happy to be proven wrong on this), therefore it ends up being broken in
lint-staged when I try it. Just run it separately, even if it's a little
slower.

See https://github.com/vuejs/language-tools/issues/686

Running tsc per-file might work with tsc-files, but that didn't work due
to Nuxt doing something clever that I don't have the patience to figure
out.