feat: skip black frames and single colors
Using ImageMagick identify's standard deviation statistic, we can skip
frames that are likely to be impossible to guess due to a lack of
features, such as black screens (from fades), single colors (from
animation transitions), etc.

This is an oft-requested feature and should prevent a lot of unfortunate
streak losses.