fix: fix image rejection and reselection
We added code in 7740237031b8a74bb9ad769cb4608bbc206f800b to
reject/regenerate frames that are uninteresting solid colors, however,
it was flawed in that it didn't change the selected episode and
timestamp when trying again. Therefore it would keep generating the same
image.

I conducted some manual testing, and it seems to work/terminate
successfully when it finds a frame that meets the threshold.