fix: fix timer updates when stats are hidden
Timer updates for the two guess timers were actually incorrectly tied to
UI updates. This resulted in them not being counted if the stats are
hidden (especially noticeable when a bunch of guesses have been made
already but the timers are zero or starting from zero), and potentially
around 0-10 ms of undercounting per frame guess.