feat: add episode button functions
It's hard to make out that it's a URL, so we add an explicit button that
shows that you can watch the episode. Also add a button to copy the
episode name and timestamp, because it's harder to copy when it's a
link, and a link to TMDB in case someone wants to edit the synopsis.

To reduce clutter these only show up on the correct answer. We can
change this in the future if needed.