feat: add explicit save button
This button saves the current frame. It has the same effect as right
clicking and saving the image, but explicitly having this button is
better because a disturbing number of people share a screenshot of the
whole page when they just want to share the current frame. It is also a
lot easier (one click) than right clicking, selecting the save option,
and then clicking the save button in the file picker dialog.