feat: show number shortcut modifier by browser
We said "Mod-N" previously for these hints, because both Ctrl and Alt
are accepted if the browser doesn't intercept them, but "Ctrl-N" and
"Alt-N" would be a lot more obvious. On Firefox, users should use Ctrl-N
because Alt-N is bound to switch to tab N, and vice versa for all other
browsers.