feat: load frames and answers from prior runs
Persist frame state and answers to disk, and be able to scan over these
datastores to repopulate the frame queue.

This will make it a lot less stressful to restart the server, because we
won't immediately start with a huge backlog. Answers being preserved
between restarts is also a nice bonus.