feat: make show data config downloadable in About
Under the GPL, we need to make a reasonable effort to provide all
resources necessary for running the site to users. While the user can
get the data from TMDB, there is some data like timings that aren't on
TMDB, and making the config downloadable just makes it easiest for
someone to set up an instance for the same show.