fix: fix save button by tweaking image lifetimes
The save button was broken in Chrome because, despite the fact that we
set "immutable" in the "Cache-Control" header, Chrome still attempts to
redownload the image and not use its cached version, possibly because it
takes the download attribute on an anchor too seriously.

We change image lifetimes on browsers other than Firefox so that images
are cleaned up when the next image is requested. This means the save
button will work as long as the image isn't expired. On Firefox, images
are cleaned up on load like before.