feat: add basic run verification
This allows the user to create a "verified run" where all frames
generated and answers given are recorded on the server. This record can
then be downloaded and presented for verification purposes.

Runs considered "important" (currently just ones that meet a certain
threshold of history entries) are archived in a separate datastore on
expiry.