fix: stop deleting answer/image when tabbing away
The visibilitychange event triggers when tabbing away, and is
potentially the last event observed before the user closes the app, and
is normally good for saving state before the user exits.  However, it
also happens when tabbing away or switching apps, which makes it too
eager to clean up the answer and image.