feat: option to name downloaded file after answer
Add an option that names the file in a "S12E34T5h12m34.567s.webp"
format. This is useful for saving the episode and timestamp information
with the file.

It is not the default behavior because this is also a spoiler when the
image is shared, if the filename is included.